Compute the kappa matrix for enrichment results

Compute the kappa matrix for enrichment results, as a measure of overlap

create_kappa_matrix(res_enrich, n_gs = nrow(res_enrich), gs_ids = NULL)

Arguments

res_enrich
A data.frame object, storing the result of the functional enrichment analysis. See more in the main function, GeneTonic(), to see the formatting requirements.

n_gs
Integer value, corresponding to the maximal number of gene sets to be included (from the top ranked ones). Defaults to the number of rows of res_enrich

gs_ids
Character vector, containing a subset of gs_id as they are available in res_enrich. Lists the gene sets to be included, additionally to the ones specified via n_gs. Defaults to NULL.

Value

A matrix with the kappa scores between gene sets

See also

Examples

# res_enrich object
data(res_enrich_macrophage, package = "GeneTonic")
res_enrich <- shake_topGOtableResult(topgoDE_macrophage_IFNg_vs_naive)

#> Found 500 gene sets in `topGOtableResult` object.

#> Converting for usage in GeneTonic...

kmat <- create_kappa_matrix(res_enrich[1:200,])
dim(kmat)

#> [1] 200 200