Simplify results from functional enrichment analysis

Simplify results from functional enrichment analysis, removing genesets that are redundant to enhance interpretation of the results

gs_simplify(res_enrich, gs_overlap = 0.75)

Arguments

res_enrich
A data.frame object, storing the result of the functional enrichment analysis. See more in the main function, GeneTonic(), to check the formatting requirements (a minimal set of columns should be present).

gs_overlap
Numeric value, which defines the threshold for removing terms that present an overlap greater than the specified value. Changing its value can control the granularity of how redundant terms are removed from the original res_enrich for the next steps, e.g. plotting this via gs_volcano()

Value

A data.frame with a subset of the original gene sets

See also

Examples

# res_enrich object
data(res_enrich_macrophage, package = "GeneTonic")
res_enrich <- shake_topGOtableResult(topgoDE_macrophage_IFNg_vs_naive)

#> Found 500 gene sets in `topGOtableResult` object.

#> Converting for usage in GeneTonic...

dim(res_enrich)

#> [1] 500   7

res_enrich_simplified <- gs_simplify(res_enrich)
dim(res_enrich_simplified)

#> [1] 143   7

# and then use this further for all other functions expecting a res_enrich